Moved all classes and functions not only related to 'SkeletonBuilder' class to file 'common.py'.
