+++ lionelmssq/common.py
@@ -2,6 +2,8 @@ from enum import Enum
 import re
 from typing import Any, Set
 
+from lionelmssq.mass_explanation import explain_mass
+
 MILP_QUASI_ONE_THRESHOLD = 0.9
 
 
@@ -32,3 +34,50 @@ def get_singleton_set_item(set_: Set[Any]) -> Any:
     if len(set_) != 1:
         raise ValueError(f"Expected a set with one item, got {set_}")
     return next(iter(set_))
+
+
+class Explanation:
+    def __init__(self, *nucleosides):
+        self.nucleosides = tuple(sorted(nucleosides))
+
+    def __iter__(self):
+        yield from self.nucleosides
+
+    def __len__(self):
+        return len(self.nucleosides)
+
+    def __repr__(self):
+        return f"{{{','.join(self.nucleosides)}}}"
+
+
+def calculate_diff_errors(mass1, mass2, threshold) -> float:
+    retval = threshold * ((mass1**2 + mass2**2) ** 0.5) / abs(mass1 - mass2)
+    # Constrain the maximum relative error to 1!
+    # For mass difference very close to zero, the relative error can be very high!
+    if retval > 1:
+        retval = 1.0
+    return retval
+
+
+def calculate_diff_dp(diff, threshold, modification_rate, seq_len, dp_table):
+    # TODO: Add support for other breakages than 'c/y_c/y'
+    explanation_list = list(
+        [
+            entry
+            for entry in explain_mass(
+                diff,
+                dp_table=dp_table,
+                seq_len=seq_len,
+                max_modifications=round(modification_rate * seq_len),
+                threshold=threshold,
+            )
+            if entry.breakage == "c/y_c/y"
+        ][0].explanations
+    )
+    if len(explanation_list) > 0:
+        retval = [
+            Explanation(*explanation_list[i]) for i in range(len(explanation_list))
+        ]
+    else:
+        retval = []
+    return retval
